toxic_comment/requirements.txt
absl-py==0.8.0
argh==0.26.2
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
docker-pycreds==0.4.0
en-core-web-sm==2.1.0
entrypoints==0.3
fire==0.2.1
gast==0.2.2
gitdb2==2.0.5
GitPython==3.0.2
google-pasta==0.1.7
gql==0.1.0
graphql-core==2.2.1
grpcio==1.23.0
h5py==2.9.0
idna==2.8
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kaggle==1.5.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.1
nvidia-ml-py3==7.352.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.3
ptyprocess==0.6.0
Pygments==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-slugify==3.0.3
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.4
requests==2.22.0
Rx==1.6.1
Send2Trash==1.5.0
sentry-sdk==0.11.2
shortuuid==0.5.0
six==1.12.0
smmap2==2.0.5
spacy==2.1.8
srsly==0.1.0
subprocess32==3.5.4
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thinc==7.0.8
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.24.3
wandb==0.8.9
wasabi==0.2.2
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
